Reorged frozen tests
