+++ pytest.ini
@@ -7,3 +7,4 @@ addopts =
     --doctest-modules credoai
     -p tests.fixtures.artifacts
     -p tests.fixtures.datasets
+    -p tests.fixtures.frozen_tests
