+++ tests/conftest.py
@@ -1,4 +1,3 @@
-import pickle
 import os
 import base64
 import json
@@ -8,7 +7,6 @@ from connect.governance import Governance
 from pandas import DataFrame
 from pytest import fixture
 
-from credoai.artifacts import ClassificationModel, TabularData
 from credoai.lens import Lens
 from connect.governance.credo_api_client import CredoApiConfig
 
@@ -155,74 +153,28 @@ def init_lens_regression(
     return my_pipeline, temp_file, gov
 
 
-################################################
-############ Frozen fixtures ###################
-################################################
-
-# with open(
-#         "tests/frozen_ml_tests/frozen_results/binary/pipeline_info.pkl", "rb"
-#     ) as f:
-#         pipeline_info = pickle.load(f)
-
-#     metrics = pipeline_info["metrics"]
-#     assessments = pipeline_info["assessments"]
-
-#     pipeline = []
-#     for assessment in assessments:
-#         pipeline.append(
-#             (
-#                 string2evaluator(assessment)(metrics),
-#                 assessment + " Assessment",
-#             )
-#         )
-
-
-@fixture(scope="session")
-def frozen_classifier():
-    # Load frozen classifier and wrap as a Credo Model
-    with open("tests/frozen_ml_tests/frozen_models/loan_binary_clf.pkl", "rb") as f:
-        clf = pickle.load(f)
-
-    return ClassificationModel("binary_clf", clf)
-
-
-@fixture(scope="session")
-def frozen_validation_data():
-    # Load frozen validation data and wrap as Credo Data
-    with open(
-        "tests/frozen_ml_tests/frozen_data/binary/loan_validation.pkl", "rb"
-    ) as f:
-        val_data = pickle.load(f)
-
-    return TabularData(
-        name=val_data["name"],
-        X=val_data["val_features"],
-        y=val_data["val_labels"],
-        sensitive_features=val_data["sensitive_features"],
-    )
-
-
-# @fixture(scope="session")
-# def frozen_training_data():
-#     # Load frozen training data and wrap as Credo Data
-#     with open("tests/frozen_ml_tests/frozen_data/binary/loan_train.pkl", "rb") as f:
-#         train_data = pickle.load(f)
-
-#     return TabularData(
-#         name=train_data["name"],
-#         X=train_data["train_features"],
-#         y=train_data["train_labels"],
-#         sensitive_features=train_data["sensitive_features"],
-#     )
+########## Integration tests config ############
 
 
 @fixture(scope="session")
 def config_path_in():
+    """
+    Retrieves config path env variable.
+
+    Not necessary in local mode as long as .config is in
+    the expected location.
+    """
     return os.getenv("CREDOAI_LENS_CONFIG_PATH", None)
 
 
 @fixture(scope="session")
 def api_config_in():
+    """
+    Retrieves config path env variable, in json format.
+
+    Not necessary in local mode as long as .config is in
+    the expected location.
+    """
     b64_config = os.getenv("CREDOAI_LENS_CONFIG_JSON_B64", None)
     if b64_config:
         return CredoApiConfig(**json.loads(base64.b64decode(b64_config)))
@@ -231,4 +183,9 @@ def api_config_in():
 
 @fixture(scope="session")
 def assessment_plan_url_in():
+    """
+    Retrieves the assessment plan designated for this test.
+
+    This is necessary also locally in order for the test to pass.
+    """
     return os.getenv("CREDOAI_LENS_PLAN_URL", None)
